I fixed the bug that was causing panels in the world gui to be rendered incorrectly as well as the scaling bug in member function SetVertices of class SnazzCraft::Panel. I also updated the B font in B.txt.

+++ SnazzCraft/includes/snazzcraft-engine/gui/gui.hpp
@@ -22,7 +22,7 @@ namespace SnazzCraft
 
         ~GUI();
 
-        virtual void Resize(uint32_t Width, uint32_t Height);
+        virtual void Resize(float Width, float Height);
 
         inline void Draw() const
         {
